id(PK)값 타입 long 통일, 오류 수정

+++ src/main/java/com/msp/membership/repository/FollowRepository.java
@@ -19,7 +19,7 @@ public interface FollowRepository extends JpaRepository<Follow, Long> {
         @Query("delete from Follow where toUserId = :toUserId and fromUserId = :fromUserId")
         void deleteFollow(@Param("toUserId") Long toUserId, @Param("fromUserId") Long fromUserId);
 
-        int countByFromUserIdAndToUserId(Long id, String userid);
+        Long countByFromUserIdAndToUserId(Long id, String userid);
 
         @Modifying
         @Transactional
